# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/users/login": {
            "post": {
                "description": "Authenticate user, set session cookie and return JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "login": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: string, data: {token: string}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "error: message",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "error: неверный логин или пароль",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "error: message",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/users/logout": {
            "post": {
                "description": "Clear session cookie and remove JWT from Redis",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "message: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "error: message",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/users/profile": {
            "get": {
                "description": "Get profile of the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.User"
                        }
                    },
                    "401": {
                        "description": "error: message",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "error: message",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "description": "Update profile of the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "description": "User info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: string",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "error: message",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "error: message",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "error: message",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/users/register": {
            "post": {
                "description": "Register a new user with login and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "data: registered user",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "error: message",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "error: message",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ds.User": {
            "type": "object",
            "properties": {
                "cargoWeight": {
                    "type": "number"
                },
                "contacts": {
                    "type": "string"
                },
                "containers20ftCount": {
                    "type": "integer"
                },
                "containers40ftCount": {
                    "type": "integer"
                },
                "fio": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "description": "\"guest\" | \"creator\" | \"moderator\"",
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        }
    }
}